+++ src/components/common/Navbar/NavbarSlider.tsx
@@ -1,6 +1,5 @@
 import { Suspense } from 'react'
-import {
-  ButtonIcon,
+import Button, {
   ButtonSignIn,
   ButtonSignInFallback,
 } from 'src/components/ui/Button'
@@ -40,7 +39,9 @@ function NavbarSlider() {
           <Logo />
         </Link>
 
-        <ButtonIcon
+        <Button
+          variant="tertiary"
+          data-fs-button-icon
           data-fs-navbar-slider-button
           aria-label="Close Menu"
           icon={<Icon name="X" width={32} height={32} />}